requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
seaborn
PyPDF2
python-docx
wordcloud